Fix tici pipeline (#4111)
When rerun test with cache, tiflash link has a bug, after
https://github.com/PingCAP-QE/ci/pull/4108/files merged

```
+ mv tiflash tiflash_dir
mv: 'tiflash' and 'tiflash_dir/tiflash' are the same file
script returned exit code 1
```

+++ pipelines/pingcap/tidb/latest/pull_integration_tici_test/pipeline.groovy
@@ -64,8 +64,13 @@ pipeline {
                                         """
                                     }
                                     sh '''
-                                        mv tiflash tiflash_dir
-                                        ln -s `pwd`/tiflash_dir/tiflash tiflash
+                                        if [[ -d tiflash && ! -L tiflash ]]; then
+                                            rm -rf tiflash_dir
+                                            mv tiflash tiflash_dir
+                                        fi
+                                        if [[ -f tiflash_dir/tiflash ]]; then
+                                            ln -sfn "$(pwd)/tiflash_dir/tiflash" tiflash
+                                        fi
                                         ls -alh .
                                         ./tikv-server -V
                                         ./pd-server -V
